Mark `FEATURE_PROTO3_OPTIONAL` as supported
We don't need to do anything to support this option, as we don't rely on
any specific field options in this generator, therefore we support his
feature by default.

This avoids warnings being emitted during generation for protos which
use `optional` fields.

This requires upgrading our dependency on `protobuf`, as `0.7.1` does
not define the `supported_features` field.